% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{Denoise}
\alias{Denoise}
\title{Applies nonlinear tree-structured wavelet thresholding based on cross-validation.}
\usage{
Denoise(P = NULL, D, alpha = 0.75, lam = NULL, return = c("P", "wt"))
}
\arguments{
\item{P}{list, signal.}
}
\value{
Returns a list containing the wavelet denoised signal..
}
\description{
Applies nonlinear tree-structured wavelet thresholding based on cross-validation.
}

