% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{E_basis}
\alias{E_basis}
\title{Orthonormal basis for space of Hermitian matrices.}
\usage{
E_basis(d)
}
\arguments{
\item{d}{the row-dimension of the matrix elements.}
}
\value{
\code{E_basis(d)} returns a \eqn{(d, d, d^2)}-dimensional array, containing the \eqn{d^2} orthonormal basis elements
 (i.e. \eqn{d x d}-dimensional matrices).
}
\description{
\code{E_basis} constructs an orthonormal basis (with respect to the Frobenius norm) of matrix elements
for the real-valued vector space of Hermitian matrices.
}
\examples{
 E_basis(3)

}

