% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{E_coeff}
\alias{E_coeff}
\title{Orthonormal basis decomposition.}
\usage{
E_coeff(H, E)
}
\arguments{
\item{H}{a \eqn{(d x d)}-dimensional Hermitian matrix.}

\item{E}{an array of orthonormal \eqn{(d x d)}-dimensional basis elements obtained by \code{E_basis(d)}.}
}
\description{
\code{E_coeff} computes the \eqn{d^2} real-valued basis coefficients with respect to an orthonormal
(in terms of the Frobenius metric) basis of the real vector space of \eqn{(d x d)}-dimensional
Hermitian matrices.
}
\examples{
 E <- E_basis(3)
 h <- matrix(complex(real = rnorm(9), imaginary = rnorm(9)), nrow = 3)
 H <- t(Conj(h)) \%*\% h
 E_coeff(H, E)

}
\seealso{
\code{\link{E_coeff_inv}}, \code{\link{E_basis}}
}

