% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{E_coeff_inv}
\alias{E_coeff_inv}
\title{Orthonormal basis recomposition.}
\usage{
E_coeff_inv(coeff, E)
}
\arguments{
\item{coeff}{a numeric vector.}

\item{E}{an array of orthonormal basis elements obtained by \code{E_basis(d)}.}
}
\description{
\code{E_coeff_inv} computes the \eqn{(d x d)}-dimensional Hermitian matrix with (real-valued) coefficients
\code{coeff} in the orthonormal basis \code{E_basis(d)}.
}
\examples{
 E <- E_basis(3)
 coeff <- rnorm(9)
 E_coeff_inv(coeff, E)

}
\seealso{
\code{\link{E_coeff}}, \code{\link{E_basis}}
}

