% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMeans.R
\name{FuzzyCmeans}
\alias{FuzzyCmeans}
\title{Fuzzy c-means clustering algorithm in wavelet domain.}
\usage{
FuzzyCmeans(P = NULL, wt = NULL, D = 2, K = 2, uptoscale = 5,
  denoise = F, m = 2, d.uptoscale = 0.1)
}
\arguments{
\item{P}{list, curves of PD matrices to be clustered.}

\item{wt}{list, wavelet representations of curves of PD matrices to be clustered.}

\item{D}{integer, number of neighbors used in wavelet transform (only necessary when \code{is.null(wt)}).}

\item{K}{integer, number of clusters.}

\item{uptoscale}{integer, number of wavelet scales to be considered in iterative clustering algorithm.}

\item{m}{numeric, fuzziness parameter (\code{m > 1}).}

\item{d.uptoscale}{numeric, if fraction of components of wavelet coefficients at scale \code{j} is smaller than
\code{d.uptoscale} the iterative clustering algorithm terminates at scale \code{j}.}

\item{anytime}{logical, output cluster assigments (yes/no) after each step in iterative clustering algorithm.}

\item{alpha}{numeric, hyperparameter for weighting function (cluster preference) in iterative clustering algorithm.}

\item{tau}{numeric, hyperparameter for weighting function (cluster preference) in iterative clustering algorithm.}
}
\value{
Returns an integer sequence with cluster assignments.
}
\description{
Fuzzy c-means clustering algorithm in wavelet domain.
}

