% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{Impute_man}
\alias{Impute_man}
\title{Impute midpoints at a finer scale (internal use only).}
\usage{
Impute_man(M_j0, D, j0, Nw)
}
\arguments{
\item{M_j0}{list of midpoints.}

\item{D}{integer.}

\item{j0}{integer.}

\item{Nw}{list of weights.}

\item{E}{list of basis elements.}
}
\value{
Returns a list of imputed midpoints at the next finer scale \code{j0 + 1}.
}
\description{
Impute midpoints at a finer scale (internal use only).
}

