% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invwavtrans.R
\name{InvWavTrans}
\alias{InvWavTrans}
\title{Computes MI inverse wavelet transform.}
\usage{
InvWavTrans(wt, D, start = 0, end = pi)
}
\arguments{
\item{wt}{list, pyramid of wavelet coefficients and midpoints on scale \eqn{j_0}.}

\item{D}{integer, number of neighbors included in MI scheme.}

\item{start}{numeric, left-boundary of frequency range, defaults to \code{0}.}

\item{end}{numeric, right-boundary of frequency range, defaults to \code{pi}.}
}
\value{
Returns a reconstruction of the curve of Hermitian PD matrices in the form
 of a list with values:
 \itemize{
   \item \code{t} the frequency points of the sampled curve.
   \item \code{p} an array of Hermitian PD matrices.
 }
 If \code{coh == TRUE} returns the squared coherence of the reconstructed
 curve of Hermitian PD matrices.
}
\description{
Computes MI inverse wavelet transform.
}

