% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invwavtrans.R
\name{InvWavTransf}
\alias{InvWavTransf}
\title{MI inverse wavelet transform.}
\usage{
InvWavTransf(D, order = 5)
}
\arguments{
\item{D}{a list of wavelet coefficients and coarse-scale midpoints as obtained from
the function \code{WavTransf}.}

\item{order}{an odd integer taking values between 1 and 9 corresponding to the
order of the MI refinement scheme.}
}
\value{
Returns an integer sequence with cluster assignments.
}
\description{
\code{InvWavTransf} computes the \emph{midpoint-interpolation} (MI) inverse wavelet
transform of a pyramid of Hermitian matrix-valued wavelet coefficients combined with an
array of coarse-scale Hermitian positive-definite midpoints as described in (Chau and von Sachs, 2017).
}
\examples{
m <- function(){
 X <- matrix(complex(real=rnorm(9), imaginary=rnorm(9)), nrow=3)
 t(Conj(X)) \%*\% X
}
M <- replicate(100, m())
z <- rnorm(100)
w <- abs(z)/sum(abs(z))
KarchMean(M, w)

}
\references{
Chau, J. and von Sachs, R. \emph{Positive-definite multivariate spectral
estimation: a geometric wavelet approach}. (Submitted)
}
\seealso{
\code{\link{WavTransf}}
}

