% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{pdDepth}
\alias{pdDepth}
\title{Zonoid and Geodesic Distance Depth for HPD matrices}
\usage{
pdDepth(y = NULL, X, method = c("zonoid", "gdd"))
}
\description{
Zonoid and Geodesic Distance Depth for HPD matrices
}

