% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{pdMean}
\alias{pdMean}
\title{Weighted Karcher mean of HPD matrices}
\usage{
pdMean(M, w, metric = "Riemannian", grad_desc = F, maxit = 1000, reltol)
}
\arguments{
\item{M}{a \eqn{(d,d,S)}-dimensional array of Hermitian PD matrices.}

\item{w}{an \eqn{S}-dimensional nonnegative weight vector, such that \code{sum(w) = 1}.}

\item{metric}{the distance measure, one of \code{'Riemannian'}, \code{'logEuclidean'},
\code{'Cholesky'}, \code{'Euclidean'} or \code{'rootEuclidean'}. Defaults to \code{'Riemannian'}.}

\item{grad_desc}{if \code{metric == "Riemannian"}, a logical value indicating if the
gradient descent algorithm should be used, defaults to \code{FALSE}.}

\item{maxit}{maximum number of iterations in gradient descent algorithm, only used if
\code{isTRUE(grad_desc & metric == "Riemannian")}. Defaults to \code{1000}}

\item{reltol}{optional tolerance parameter in gradient descent algorithm, only used if
\code{isTRUE(grad_desc & metric == "Riemannian")}. Defaults to \code{1E-10}.}
}
\description{
\code{pdMean} calculates an (approximate) weighted Karcher or Frechet mean of \eqn{S} different
\eqn{(d \times d)}-dimensional Hermitian PD matrices intrinsic to a specific metric. For the
affine-invariant Riemannian metric, the weighted Karcher mean is either approximated via
the fast recursive algorithm in (Ho et al., 2016) or computed via the slower, but more accurate,
gradient descent algorithm in (Pennec, 2006). By default, the unweighted Karcher mean is computed.
}
\examples{
m <- function(){
 X <- matrix(complex(real=rnorm(9), imaginary=rnorm(9)), nrow=3)
 t(Conj(X)) \%*\% X
}
M <- replicate(100, m())
z <- rnorm(100)
w <- abs(z)/sum(abs(z))
pdMean(M, w)

}
\references{
Chau, J. and von Sachs, R. (2017). \emph{Intrinsic wavelet regression for curves of
Hermitian positive definite matrices}. Available at \url{http://arxiv.org/abs/1701.03314}.

Pennec, X. (2006). Intrinsic statistics on Riemannian manifolds: Basic tools for geometric
measurements. \emph{Journal of Mathematical Imaging and Vision} 25(1), 127-154.
}
\seealso{
\code{\link{Mid}}, \code{\link{pdMedian}}
}
