% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{pdMedian}
\alias{pdMedian}
\title{Weighted intrinsic median of HPD matrices}
\usage{
pdMedian(M, w, max_iter = 1000, tol, ...)
}
\arguments{
\item{M}{a \eqn{(d,d,S)}-dimensional array of Hermitian PD matrices.}

\item{w}{an \eqn{S}-dimensional nonnegative weight vector, such that \code{sum(w) = 1}.}

\item{max_iter}{maximum number of iterations in the steepest descent algorithm.}

\item{tol}{optional tolerance parameter in the steepest descent algorithm, defaults to \code{.Machine$double.eps}.}

\item{...}{additional arguments for internal usage.}
}
\description{
\code{pdMedian} calculates an (approximate) weighted intrinsic of \eqn{S} different
\eqn{(d \times d)}-dimensional Hermitian PD matrices based on the Riemannian metric by the
intrinsic Weiszfeld algorithm in (Fletcher et al., 2009). By default, the unweighted intrinsic median is computed.
}
\examples{
m <- function(){
 X <- matrix(complex(real=rnorm(9), imaginary=rnorm(9)), nrow=3)
 t(Conj(X)) \%*\% X
}
M <- replicate(100, m())
z <- rnorm(100)
w <- abs(z)/sum(abs(z))
pdMedian(M, w)

}
\references{
Fletcher, P.T., Venkatasubramanian, S. and Joshi, S. (2009). The geometric median
on Riemannian manifolds with application to robust atlas estimation. \emph{NeuroImage} 45(1),
S143-S152.
}
\seealso{
\code{\link{pdMean}}
}
