% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgram.R
\name{pdPgram2D}
\alias{pdPgram2D}
\title{Tapered HPD time-varying periodogram matrix}
\usage{
pdPgram2D(X, B, method = c("DPSS", "Hermite"), bias.corr = F, time, freq)
}
\description{
Tapered HPD time-varying periodogram matrix
}
