% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranktests.R
\name{pdRankTests}
\alias{pdRankTests}
\title{Depth-rank hypothesis testing for HPD matrices}
\usage{
pdRankTests(samples, sample.sizes, depth = c("gdd", "zonoid", "spatial"),
  test = c("rank.sum", "krusk.wall", "signed.rank", "bartels", "rank.sum.fun",
  "bartels.fun"))
}
\description{
Depth-rank hypothesis testing for HPD matrices
}
