% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{pdSplineReg}
\alias{pdSplineReg}
\title{HPD cubic smoothing spline regression}
\usage{
pdSplineReg(P, f.hat0, lam, K, initial_step = 1, max.iter = 100,
  eps = 0.001, ...)
}
\description{
HPD cubic smoothing spline regression
}
