% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{rExamples1D}
\alias{rExamples1D}
\title{Several example spectral matrices}
\usage{
rExamples1D(n, d = 3, example = c("heaviSine", "bumps", "two-cats",
  "gaussian", "mix-gaussian"), return.ts = FALSE, noise = "riem-gaussian",
  noise.level = 1, df.wishart = NULL)
}
\arguments{
\item{n}{number of sampled matrices to be generated.}

\item{d}{row- (resp. column-)dimension of the generated matrices.}

\item{example}{the example HPD spectral matrix curve, one of \code{'heaviSine'}, \code{'bumps'}, \code{'two-cats'},
\code{'gaussian'} or \code{'mix-gaussian'}.}

\item{return.ts}{if \code{isTRUE(return.ts)} the function returns time series observations generated via the Cram\'er representation
based on the transfer function of the example spectral matrix and complex normal random variates.}

\item{noise}{noise distribution for the generated noisy spectral matrices, one of \code{'riem-gaussian'},
\code{'periodogram'}, \code{'log-gaussian'}, \code{'wishart'} or \code{'log-wishart'}. Additional details are given below.}

\item{noise.level}{parameter to tune the signal-to-noise ratio for the generated noisy spectral matrices.}

\item{df.wishart}{optional parameter to specify the degrees of freedom in the case of a Wishart noise distribution or the
number of DPSS taper functions in the case of generated periodogram matrices. Only used if \code{noise} is equal to \code{'periodogram'},
\code{'wishart'} or \code{'log-wishart'}. By default \code{df.wishart} is equal to the dimension \code{d} to
guarantee positive definiteness of the generated matrices.}
}
\value{
Depending on the input arguments returns a list with two or three components:
   \item{\code{f} }{ example spectral matrix, \code{f} is a \code{(d, d, length(freq))}-dimensional array, corresponding
   to a \eqn{(d \times d)}-dimensional spectral matrix at the Fourier frequencies \code{freq}.}
   \item{\code{P} }{ multitaper HPD periodogram of the generated time series, by default pre-smoothed using \code{df.wishart = d}
   DPSS tapering functions, see \code{\link{pdPgram}} for details.}
   \item{\code{ts}}{ generated \eqn{d}-dimensional time series observations, only available if \code{isTRUE(return.ts)}.}
}
\description{
\code{rExamples1D()} generates stationary time series observations from several example HPD spectral matrices
for testing and simulation purposes. For more details, we refer to the simulation studies in (Chau and von Sachs, 2017).
}
\details{
The examples include: (i) a \eqn{(3 \times 3)} heaviSine HPD spectral matrix consisting of smooth sinosoids with a break,
(ii) a \eqn{(3 \times 3)} bumps HPD spectral matrix containing peaks and bumps of various smoothness degrees, (iii) a
\eqn{(3 \times 3)} two-cats HPD spectral matrix visualizing the contour of two side-by-side cats, with inhomogeneous
smoothness across frequency, and (iv) a \eqn{(2 \times 2)} Gaussian HPD spectral matrix consisting of smooth random Gaussian
functions. The time series observations are generated via the Cram\'er representation based on the transfer function of
the example spectral matrix and complex normal random variates.
}
\examples{
example <- rExamples1D(100, example = "bumps", return.ts = TRUE)
ts.plot(Re(example$ts)) # plot generated time series

}
\references{
Chau, J. and von Sachs, R. (2017). \emph{Positive definite multivariate spectral
estimation: a geometric wavelet approach}. Available at \url{http://arxiv.org/abs/1701.03314}.
}
\seealso{
\code{\link{pdPgram}}
}
