% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{SpecClust}
\alias{SpecClust}
\title{Wavelet-based clustering of multivariate spectra.}
\usage{
SpecClust(P = NULL, wt = NULL, D = 2, K = 2, jmax = NULL,
  denoise = F, m = 2, d.jmax = 0.1)
}
\arguments{
\item{P}{list, curves of PD matrices to be clustered.}

\item{wt}{list, wavelet representations of curves of PD matrices to be clustered.}

\item{D}{integer, number of neighbors used in wavelet transform (only necessary when \code{is.null(wt)}).}

\item{K}{integer, number of clusters.}

\item{m}{numeric, fuzziness parameter (\code{m > 1}).}

\item{uptoscale}{integer, number of wavelet scales to be considered in iterative clustering algorithm.}

\item{anytime}{logical, output cluster assigments (yes/no) after each step in iterative clustering algorithm.}

\item{alpha}{numeric, hyperparameter for weighting function (cluster preference) in iterative clustering algorithm.}

\item{tau}{numeric, hyperparameter for weighting function (cluster preference) in iterative clustering algorithm.}

\item{d.uptoscale}{numeric, if fraction of components of wavelet coefficients at scale \code{j} is smaller than
\code{d.uptoscale} the iterative clustering algorithm terminates at scale \code{j}.}
}
\value{
Returns an integer sequence with cluster assignments.
}
\description{
\code{SpecClust} performs clustering of multivariate spectral matrices via a two-step fuzzy
clustering algorithm in the manifold wavelet domain as detailed in (Chau and von Sachs, 2017).
}
\details{
This is a two-step clustering algorithm.
}
\examples{
m <- function(){
 X <- matrix(complex(real=rnorm(9), imaginary=rnorm(9)), nrow=3)
 t(Conj(X)) \%*\% X
}
M <- replicate(100, m())
z <- rnorm(100)
w <- abs(z)/sum(abs(z))
KarchMean(M, w)

}
\references{
Chau, J. and von Sachs, R. \emph{Positive-definite multivariate spectral
estimation: a geometric wavelet approach}. (Submitted)
}

