% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavtrans.R
\name{WavTrans}
\alias{WavTrans}
\title{Computes MI forward wavelet transform of a curve of Hermitian PD matrices.}
\usage{
WavTrans(P = NULL, D, M = NULL, uptoscale = NULL)
}
\arguments{
\item{P}{list, sampled curve of Hermitian PD matrices, defaults to \code{NULL}.}

\item{D}{integer, number of neighbors included in MI scheme.}

\item{M}{list, midpoint pyramid, defaults to \code{NULL}.}
}
\value{
Returns a list with values:
 \itemize{
   \item \code{wt} pyramid of wavelet coefficients (Hermitian matrices).
   \item \code{M} pyramid of midpoints.
 }
If \code{M == NULL} is \code{FALSE}, the function computes the pyramid of wavelet
coefficients directly from the given midpoint pyramid \code{M}.
}
\description{
Computes MI forward wavelet transform of a curve of Hermitian PD matrices.
}

