% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavtrans.R
\name{WavTransf}
\alias{WavTransf}
\title{MI forward wavelet transform.}
\usage{
WavTransf(P, order = 5, jmax)
}
\arguments{
\item{P}{a (\eqn{d,d,m})-dimensional array, with \eqn{m} a dyadic number.}

\item{order}{an odd integer taking values between 1 and 9 corresponding to the
order of the MI refinement scheme.}

\item{jmax}{the maximum scale at which the wavelet coefficients are computed, such
that \eqn{jmax < log2(m)}.}
}
\value{
Returns an integer sequence with cluster assignments.
}
\description{
\code{WavTransf} computes the \emph{midpoint-interpolation} (MI) forward wavelet
transform of a curve of length \eqn{m} of (\eqn{d x d})-dimensional HPD matrices
(e.g. a multivariate spectrum) as described in (Chau and von Sachs, 2017).
}
\examples{
m <- function(){
 X <- matrix(complex(real=rnorm(9), imaginary=rnorm(9)), nrow=3)
 t(Conj(X)) \%*\% X
}
M <- replicate(100, m())
z <- rnorm(100)
w <- abs(z)/sum(abs(z))
KarchMean(M, w)

}
\references{
Chau, J. and von Sachs, R. \emph{Positive-definite multivariate spectral
estimation: a geometric wavelet approach}. (Submitted)
}
\seealso{
\code{\link{InvWavTransf}}
}

