% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{pdCART}
\alias{pdCART}
\title{Tree-structured wavelet-thresholded 2D spectral estimator}
\usage{
pdCART(D, D_white, tree = T, lam = NA, alpha = 1, periodic = T)
}
\description{
Tree-structured wavelet-thresholded 2D spectral estimator
}
