% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranktests.R
\name{pdRankTests}
\alias{pdRankTests}
\title{Rank-based hypothesis tests for HPD matrices}
\usage{
pdRankTests(data, sample.sizes, depth = c("gdd", "zonoid", "spatial"),
  test = c("rank.sum", "krusk.wall", "signed.rank", "bartels"))
}
\arguments{
\item{data}{either a \eqn{(d,d,S)}-dimensional array corresponding to an array of pooled individual samples of Hermitian PD matrices, or a
\eqn{(d,d,n,S)}-dimensional array corresponding to an array of pooled individual samples of sequences of Hermitian PD matrices.}

\item{sample.sizes}{a numeric vector corresponding to the individual sample sizes in the argument \code{data}, such that \code{sum(sample.sizes) == S}.
For tests \code{"signed-rank"} and \code{"bartels"}, \code{sample.sizes} is not required and is automatically determined from \code{data}.}

\item{depth}{data depth measure used in the rank-based tests, one of \code{"gdd"}, \code{"zonoid"}, or \code{"spatial"} corresponding to the
geodesic distance depth, manifold zonoid depth and manifold spatial depth respectively. Defaults to \code{"gdd"} and is not required for test
\code{"signed.rank"}.}

\item{test}{rank-based hypothesis testing procedure, one of \code{"rank.sum"}, \code{"krusk.wall"}, \code{"signed.rank"}, \code{"bartels"} explained
in the Details section below.}
}
\value{
The function returns a list with three components:
\item{p.value }{p-value of the test}
\item{statistic }{test statistic of the test}
\item{null.distr }{the distribution of the test statistic under the null hypothesis}
}
\description{
\code{pdRankTests} performs generalized rank-based hypothesis testing for samples of Hermitian PD matrices or
samples of sequences (arrays) of Hermitian PD matrices as detailed in (Chau and von Sachs, 2017b).
}
\details{
For samples of \eqn{(d \times d)}-dimensional Hermitian PD matrices with pooled sample size \eqn{S}, the argument
\code{data} is a \eqn{(d,d,S)}-dimensional array of Hermitian PD matrices, where the individual samples are
combined along the third array dimension (using e.g. \code{\link[abind]{abind}}). For samples of sequences of
\eqn{(d \times d)}-dimensional Hermitian PD matrices with pooled sample size \eqn{S}, the argument \code{data} is a
\eqn{(d,d,n,S)}-dimensional array of sequences of Hermitian PD matrices, where the individual samples
are combined along the fourth array dimension. The argument \code{sample.sizes} specifies the sizes of the individual
samples so that \code{sum(sample.sizes) == S}. \cr
The available generalized rank-based testing procedures, specificied by the argument \code{test}, are:
\describe{
\item{\code{"rank.sum"}}{Manifold Wilcoxon rank-sum test to test for homogeneity of distributions of two independent
samples of Hermitian PD matrices or samples of sequences of Hermitian PD matrices. The usual univariate ranks are replaced by data depth
induced ranks via \code{\link{pdDepth}}.}
\item{\code{"krusk.wall"}}{Manifold Kruskall-Wallis test to test for homogeneity of distributions of more than two independent
samples of Hermitian PD matrices or samples of sequences of Hermitian PD matrices. The usual univariate ranks are replaced by data depth
induced ranks via \code{\link{pdDepth}}.}
\item{\code{"signed-rank"}}{Manifold signed-rank test to test for homogeneity of distributions of independent paired or matched samples
of Hermitian PD matrices. The manifold signed-rank test is \emph{not} based on data depth induced ranks, but on a specific difference score on the Riemannian
manifold of Hermitian PD matrices.}
\item{\code{"bartels"}}{Manifold Bartels-von Neumann test to test for randomness (i.e. exchangeability) within a single indepdendent sample of
Hermitian PD matrices or a sample of sequences of Hermitian PD matrices. The usual univariate ranks are replaced by data depth induced
ranks via \code{\link{pdDepth}}.}
}
Note in particular that the manifold signed-rank test also provides a valid test for equivalence of spectral matrices of two multivariate stationary time
series based on the Hermitian PD periodogram matrices obtained via \code{\link{pdPgram}}, see (Chau and von Sachs, 2017b) for the details.
}
\examples{
## null hypothesis is true
E <- pdSpecEst:::E_basis(2)
data <- replicate(100, Expm(diag(2), pdSpecEst:::E_coeff_inv(rnorm(4), E)))
pdRankTests(data, sample.sizes = c(50, 50), test = "rank.sum") ## homogeneity 2 samples
pdRankTests(data, sample.sizes = rep(25, 4), test = "krusk.wall") ## homogeneity 4 samples
pdRankTests(data, test = "bartels") ## randomness

## null hypothesis is false
data1 <- abind::abind(data, replicate(50, Expm(diag(2), pdSpecEst:::E_coeff_inv(0.5 * rnorm(4), E))))
pdRankTests(data1, sample.sizes = c(100, 50), test = "rank.sum")
pdRankTests(data1, sample.sizes = rep(50, 3), test = "krusk.wall")
pdRankTests(data1, test = "bartels")

## signed-rank test for equivalence of spectra
## ARMA(1,1) process: Example 11.4.1 in (Brockwell and Davis, 1991)
Phi <- array(c(0.7, 0, 0, 0.6, rep(0, 4)), dim = c(2, 2, 2))
Theta <- array(c(0.5, -0.7, 0.6, 0.8, rep(0, 4)), dim = c(2, 2, 2))
Sigma <- matrix(c(1, 0.71, 0.71, 2), nrow = 2)
pgram <- function(Sigma) pdPgram(rARMA(2^9, 2, Phi, Theta, Sigma)$X)$P

## null is true
pdRankTests(abind::abind(pgram(Sigma), pgram(Sigma)), test = "signed.rank")
## null is false
pdRankTests(abind::abind(pgram(Sigma), pgram(0.5 * Sigma)), test = "signed.rank")

}
\references{
Chau, J. and von Sachs, R. (2017b). \emph{Statistical data depth and rank-based
tests for spectral density matrices}. Available at \url{http://arxiv.org/abs/...}.

Brockwell, P.J. and Davis, R.A. (1991). \emph{Time series: Theory and Methods}. New York: Springer.
}
\seealso{
\code{\link{pdDepth}}, \code{\link{pdPgram}}
}
