% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{pdSpecEst1D}
\alias{pdSpecEst1D}
\title{Automatic tree-structured wavelet regression for curves of HPD matrices}
\usage{
pdSpecEst1D(P, order = 5, policy = c("universal", "cv"),
  metric = "Riemannian", periodic = T, alpha = 1, return = "f", ...)
}
\description{
Automatic tree-structured wavelet regression for curves of HPD matrices
}
